requirements.txt
pip~=23.2.1
wheel~=0.41.2
mysql-connector-python~=9.1.0
python-dotenv~=1.0.1